packages/stage-one: add pkg-config-lite 0.28-1
Several packages will use pkg-config to find various required and optional
dependencies. Unfortunately, this could result in those packages finding
libraries on the host system and trying to link against them, which will
not work.

To avoid this problem, install a pkg-config binary in /cross-tools and
configure it so that it will look for Pkg-config files only in $(SYSROOTDIR).